Sleep cycle

+++ REMGuide/Models/ViewModels/HomePageViewModel.cs
@@ -8,6 +8,6 @@ namespace REMGuide.Models.ViewModels
     public class HomePageViewModel
     {
         public ICollection<Theme> TopThemes { get; set; }
-        public int AvgDisruptions { get; set; }
+        public double AvgDisruptions { get; set; }
     }
 }
